* [in progress]  add OperationInfo

* -add operationInfo
- Goal: get information for current idle items and use it for ML

* - fix: calculate idletime for each capability for each interval

* Add new KPIs for amount of idle jobs
Add AI project with regression model

* Update to .NET 6.0
change target plattform x64 for ML.NET Builder

* fix bug for order pop and next order pop

* update to dotnet 6.0

* add arch

* remove target platform anyCPU

* feature: add KI prediction to popOrder
 - estimating time that order will be probably to early if released without waiting
- feature-list: total processing duration, longest processing duration, time from creation to due
improvements: add level to throughput anaylzer
bugfixes: there are no (known) bugs here - so nothing to fix :)

* notice that this version is still running on AkkaSim and has not been updating to AkkaHive

+++ Mate.Production.Core/Environment/Options/TransitionFactor.cs
@@ -3,9 +3,9 @@ using Mate.Production.Core.Environment.Abstractions;
 
 namespace Mate.Production.Core.Environment.Options
 {
-    public class TransitionFactor : Option<decimal>
+    public class TransitionFactor : Option<double>
     {
-        public TransitionFactor(decimal value)
+        public TransitionFactor(double value)
         {
             _value = value;
         }
@@ -13,7 +13,7 @@ namespace Mate.Production.Core.Environment.Options
         public TransitionFactor()
         {
             Action = (config, argument) => {
-                config.AddOption(o: new TransitionFactor(value: decimal.Parse(argument)));
+                config.AddOption(o: new TransitionFactor(value: double.Parse(argument)));
             };
         }
     }
